Nurse_Ni/requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
azure-cognitiveservices-speech==1.37.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
line-bot-sdk==3.11.0
lxml==5.2.2
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openai==1.30.4
packaging==24.0
pandas==2.2.2
parsel==1.9.1
pydantic==2.7.2
pydantic_core==2.18.3
pydub==0.25.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
w3lib==2.1.2
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4